requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
cached-property==1.3.1
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
cryptography==2.1.4
defusedxml==0.5.0
Django==1.11.10
django-appconf==1.0.2
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-rq==1.0.1
django-select2==6.0.0
django-uuslug==1.1.8
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
easy-thumbnails==2.5
enum34==1.1.6
et-xmlfile==1.0.1
idna==2.6
ipaddress==1.0.19
isodate==0.6.0
jdcal==1.3
lxml==4.1.1
Markdown==2.6.11
ntlm-auth==1.0.6
olefile==0.44
openpyxl==2.4.5
packaging==16.8
Pillow==4.0.0
psycopg2==2.7.1
pycparser==2.18
PyJWT==1.5.3
pyparsing==2.2.0
python-memcached==1.58
python-slugify==1.2.4
pytz==2018.3
redis==2.10.6
requests==2.18.4
requests-ntlm==1.1.0
requests-toolbelt==0.8.0
rq==0.10.0
six==1.10.0
sqlparse==0.2.4
Unidecode==0.4.21
urllib3==1.22
zeep==2.5.0
